requirements.txt
# Core libraries
torch>=1.7.0
torchvision>=0.8.0
ultralytics>=8.0.0  # YOLOv8 library
numpy>=1.20.0
matplotlib>=3.3.0
tqdm>=4.60.0
Pillow>=8.0.0
PyYAML>=6.0

# Data processing and visualization
pycocotools>=2.0.0  # For COCO dataset handling
opencv-python>=4.5.0
albumentations>=1.0.0
scikit-learn>=1.0.0
seaborn>=0.11.0

# Utilities
ipywidgets>=7.0.0  # For interactive widgets in notebooks
jupyterlab>=3.0.0  # Recommended for notebook development
tensorboard>=2.5.0  # For visualizing training progress